basic plugin practice
Install extension:
1) Open Google Chrome and go to chrome://extensions.
2) Enable the "Developer mode" option.
3) Click on "Load unpacked" and select the directory where your extension project is located.
4) The extension should now be loaded into Chrome. You will see the extension icon in the browser toolbar.

NOTE: only some web pages allow you to inject scripts, so the extension doesn't work on many web pages. Try it on this page (it will work):
https://www.scu.edu/academics/undergraduate-programs/

Functionality:
1) POPUP BUTTON
Click on the extension icon, and when you click the "Change Color" button in the popup, the background color of the current webpage will change to red.
Clicking again will revert the color to its original state.

2) CONTEXT MENU
Right click on screen to see context menu.
There is a parent item called "change background color" with a paintbrush icon.
The two child options are "on" and "off."
The context menu has the same functionality as the button popup.